\name{mySessionInfo}
\alias{mySessionInfo}

\title{ Wrapper around sessionInfo }
\description{
  This will produce the LaTeX output for the sessionInfo and the
  references at the end of each lab.
}
\usage{
mySessionInfo(ref=TRUE)
}

\arguments{
  \item{ref}{logical controlling whether to include references}
  }

\value{
 This function is called for its side effects
}

\author{Florian Hahne}


\keyword{misc}